+++ src/drawing.rs
@@ -1,7 +1,6 @@
 use crate::color::Color;
 use crate::drawable::{DrawType, Drawable};
 use crate::image::Image;
-use crate::image_loading::indexed::IndexedImage;
 use crate::shapes::CreateDrawable;
 use crate::text::format::TextFormat;
 use crate::text::pos::TextPos;
@@ -14,6 +13,8 @@ use graphics_shapes::polygon::Polygon;
 use graphics_shapes::rect::Rect;
 use graphics_shapes::triangle::Triangle;
 use std::mem::swap;
+use ici_files::animated::AnimatedIndexedImage;
+use ici_files::image::IndexedImage;
 
 /// Represents anything that [Graphics] can render
 pub trait Renderable<T> {
@@ -119,13 +120,34 @@ impl Graphics<'_> {
     }
 
     /// Draw an indexed image at `x`, `y`
+    #[cfg(feature="indexed_images")]
     pub fn draw_indexed_image<P: Into<Coord>>(&mut self, xy: P, image: &IndexedImage) {
         let xy = xy.into();
-        for x in 0..image.width() {
-            for y in 0..image.height() {
-                let i = x + y * image.width();
-                let color = image.colors()[image.pixels()[i] as usize];
-                self.set_pixel(x as isize + xy.x, y as isize + xy.y, color);
+        let palette = image.get_palette();
+        let (width,height) = image.size();
+        for x in 0..width {
+            for y in 0..height {
+                let i = image.get_pixel_index(x,y).unwrap();
+                let color_idx = image.get_pixel(i).unwrap() as usize;
+                let color = palette[color_idx];
+                self.set_pixel(x as isize + xy.x, y as isize + xy.y, color.into());
+            }
+        }
+    }
+
+    /// Draw an animated image at `x`, `y`
+    #[cfg(feature="indexed_images")]
+    pub fn draw_animated_image<P: Into<Coord>>(&mut self, xy: P, image: &AnimatedIndexedImage) {
+        let xy = xy.into();
+        let palette = image.get_palette();
+        let (width,height) = image.size();
+        let current_frame = image.get_current_frame_pixels();
+        for x in 0..width {
+            for y in 0..height {
+                let i = image.get_pixel_index(x,y).unwrap();
+                let color_idx = current_frame[i] as usize;
+                let color = palette[color_idx];
+                self.set_pixel(x as isize + xy.x, y as isize + xy.y, color.into());
             }
         }
     }